[crashtracker] Refactor receiver manager
