+++ crashtracker/src/collector/mod.rs
@@ -7,6 +7,7 @@ mod atomic_set;
 mod counters;
 mod crash_handler;
 mod emitters;
+mod receiver_manager;
 mod saguard;
 mod signal_handler_manager;
 mod spans;
