+++ ddcommon/src/lib.rs
@@ -25,6 +25,7 @@ pub mod hyper_migration;
 pub mod rate_limiter;
 pub mod tag;
 pub mod tracer_metadata;
+pub mod unix_utils;
 
 /// Extension trait for `Mutex` to provide a method that acquires a lock, panicking if the lock is
 /// poisoned.
